Adding html and css (without escaping data)
in this session I added html and css into my project and prepared it with partials.
now the server configuration is fine but still I need to understand it more. Also we discussed something called output buffering
we added a new method to our TemplateEngine class called resolve to be responsible of our static include of the files